allowed arbitrary files for checkin. Added docs

+++ obo-checkin.pl
@@ -5,6 +5,7 @@ use FileHandle;
 my $outdir = "terms";
 my $cmd;
 my $dry_run = 0;
+my $preserve_files = 0;
 while ($ARGV[0] =~ /^\-/) {
     my $opt = shift @ARGV;
     if ($opt eq '-h' || $opt eq '--help') {
@@ -16,6 +17,9 @@ while ($ARGV[0] =~ /^\-/) {
     }
     if ($opt eq '-n' || $opt eq '--dry-run') {
         $dry_run = 1;
+    }
+    if ($opt eq '-p' || $opt eq '--preserve-files') {
+        $preserve_files = 1;
     }
 }
 `mkdir -p $outdir`;
@@ -30,23 +34,48 @@ my %new_stanza_map = ();
 
 foreach my $id (@ids) {
     my $path = get_path($id);
-    open(F, $path) || die "no such file $path";
-    my $stanza = "";
-    while(<F>) {
-        chomp;
-        $stanza .= "$_\n";
-    }
-    close(F);
-    if ($stanza =~ /id: (\S+)/) {
-        # check id matches
-        if ($1 ne $id) {
-            die "id mismatch $1 ne $id";
+    # check if $id is a path to a file that exists
+    if ($id =~ m@[\./]@ && -e $id) {
+        open(F, $id) || die "no such file $id";
+        my @lines = <F>;
+        close(F);
+        my $uber_stanza = join("", @lines);
+        my @stanzas_in_block = split(/\n\n/, $uber_stanza);
+        foreach my $stanza (@stanzas_in_block) {
+            # trim whitespace
+            $stanza =~ s/\s+$//;
+            if (!length($stanza)) {
+                next;
+            }
+            # check if stanza has id (note that stanza is multi-line)
+            if ($stanza =~ /id:\s+(\S+)/) {
+                my $stanza_id = $1;
+                $new_stanza_map{$stanza_id} = "$stanza\n\n";
+            }
+            else {
+                die "no id found in $stanza";
+            }
         }
     }
     else {
-        die "no id found in $path";
+        open(F, $path) || die "no such file $path";
+        my $stanza = "";
+        while(<F>) {
+            chomp;
+            $stanza .= "$_\n";
+        }
+        close(F);
+        if ($stanza =~ /id: (\S+)/) {
+            # check id matches
+            if ($1 ne $id) {
+                die "id mismatch $1 ne $id";
+            }
+        }
+        else {
+            die "no id found in $path";
+        }
+        $new_stanza_map{$id} = $stanza;
     }
-    $new_stanza_map{$id} = $stanza;
 }
 
 open(W, ">$fn.tmp") || die "cannot write tp $fn.tmp";
@@ -81,7 +110,8 @@ foreach my $id (sort keys %new_stanza_map) {
     
     # Update stanza type for new stanzas
     if ($new_stanza_map{$id} =~ /\[(\w+)\]/) {
-        $stanza_type_map{$id} = $1;
+        my $s = $1;
+        $stanza_type_map{$id} = $s;
     }
     else {
         # Default to Term if type not specified
@@ -99,7 +129,10 @@ my @sorted_ids = sort {
 } keys %stanza_map;
 
 foreach my $id (@sorted_ids) {
-    print W $stanza_map{$id};
+    my $s = $stanza_map{$id};
+    # normalize line endings to strip trailing whitespace
+    $s =~ s@[\r\n]+$@\n\n@;
+    print W $s;
 }
 close(W);
 
@@ -111,14 +144,25 @@ else {
     # clear out @ids from $outdir
     foreach my $id (@ids) {
         my $path = get_path($id);
-        unlink $path;
+        if (!$preserve_files) {
+            unlink $path;
+        }
     }
 }
 
+# get the path for an id
+# the ID should be either:
+# - an ontology curie, e.g. GO:0000001, in which case the path is terms/GO_0000001.obo
+# - an OWL local name, e.g. GO_0000001, in which case the path is terms/GO_0000001.obo
+# - a file name, e.g. terms/my_terms.obo, in which case the path is terms/my_terms.obo
 sub get_path {
     my ($id) = @_;
     my $fn = "$id";
     $fn =~ s@:@_@;
+    # if the id has : or / in it and is a path to a file that exists, return it
+    if ($fn =~ m@[\./]@ && -e $fn) {
+        return $fn;
+    }
     return "$outdir/$fn.obo"
     
 }
@@ -141,13 +185,16 @@ sub usage {
     my $sn = scriptname();
 
     <<EOM;
-$sn [-s chunksize] [-x COMMAND \;] OBO-FILES
+$sn  OBO-FILE [ -d TERM-DIR ] TERM1 TERM2 ...
 
-Splits obo file into stanzas
+Checks in obo files from TERM-DIR into the OBO-FILE
 
 Example:
 
-$sn -d terms chebi.obo 
+$sn src/ontology/foo-edit.obo FOO:0000087 FOO:0000081
+
+This will check in the FOO:0000087 and FOO:0000081 terms from the terms directory
+into the foo-edit.obo file.
 
 EOM
 }
